fix build after renaming
