+++ src/components/Features/ProfileCardFeature.vue
@@ -1,7 +1,8 @@
 <script setup>
 import { UserConfig } from "@data/website";
-import GlassCard from "./GlassCard.vue";
-import SocialMediaShortcuts from "./SocialMediaShortcuts.vue";
+
+import GlassCard from "@/components/GlassCard.vue";
+import SocialMediaShortcuts from "@/components/SocialMediaShortcuts.vue";
 import AsyncImage from "@/components/AsyncImage.vue";
 </script>
